Minor fix: force hexdec to be int

+++ src/Soluble/Japha/Bridge/Driver/Pjb62/Client.php
@@ -366,11 +366,11 @@ class Client
     /**
      * @param string $str
      *
-     * @return int|float
+     * @return int
      */
     protected function getExact($str)
     {
-        return hexdec($str);
+        return (int) hexdec($str);
     }
 
     /**
